backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases[postgresql]
asyncpg
python-dotenv
alembic
httpx==0.24.0
